+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add adapter-internal interpolation of boundary functions for 2D and 3D cases. [#55](https://github.com/precice/fenicsx-adapter/pull/55)
 * Add MPI support. [#55](https://github.com/precice/fenicsx-adapter/pull/55)
+* Added functions which wrap profiling API [#56](https://github.com/precice/fenicsx-adapter/pull/56)
 * Remove deepcopy operations in `read_data` as they are not performance efficient [#52](https://github.com/precice/fenicsx-adapter/pull/52)
 * Support 3D coupling. [#51](https://github.com/precice/fenicsx-adapter/pull/51)
 * Support JIT-mapping. [#48](https://github.com/precice/fenicsx-adapter/pull/48)